extract workspace activation logic
